Create ReverseNumber.java
This code prompts the user to enter a number, and then uses a loop to reverse the digits of the number. The reverse method first initializes the reversedNum variable to 0. It then enters a loop which continues until num is 0. Inside the loop, the reversedNum variable is updated to be equal to itself multiplied by 10 plus the last digit of num. The num variable is then updated to be equal to itself divided by 10. When the loop ends, the reversedNum variable contains the reversed number.